requirements.txt
requests==2.32.3 
duckdb==1.1.2
pandas==2.2.3
google-api-python-client==2.149.0
google-auth-oauthlib==1.2.1
pyarrow==17.0.0